Duh! Found the bug.
Bug was actually cause by a logical error in #Schedule::loop.
Namely, the iteration has to account for the fact that elements are intentionally deleted from the vector in the loop and potentially added at any call of #Event::tryExecute. As such, iterators can't be used. Iteration was switched to indices.